fix(manager): keep attention ordering browser-safe
