+++ src/process_diagnostics.ts
@@ -4,6 +4,7 @@ import type {RuntimeConfig} from './types.js';
 import {SystemInfo} from './effect/system.js';
 import {readLiveStandaloneProcessLeases} from './standalone_process_lease.js';
 import {sha256HexSync} from './crypto/sha256.js';
+import {orderThreadnoteProcessesByAttention} from './process_attention.js';
 import {withoutTelemetrySessionEnvironment} from './telemetry/session.js';
 
 const PROCESS_DIAGNOSTICS_SCHEMA_VERSION = 1;
@@ -70,49 +71,6 @@ export interface ManageableThreadnoteProcessDiagnostics {
   readonly processes: readonly ManageableThreadnoteProcessDiagnostic[];
   readonly schemaVersion: typeof PROCESS_DIAGNOSTICS_SCHEMA_VERSION;
   readonly truncated: boolean;
-}
-
-export function orderThreadnoteProcessesByAttention<T extends ThreadnoteProcessDiagnostic>(
-  processes: readonly T[],
-): readonly T[] {
-  return processes
-    .map((process, inputIndex) => ({inputIndex, process}))
-    .sort(
-      (left, right) =>
-        processAttentionRank(left.process) - processAttentionRank(right.process) ||
-        left.process.startedAt.localeCompare(right.process.startedAt) ||
-        left.process.processId - right.process.processId ||
-        left.inputIndex - right.inputIndex,
-    )
-    .map(entry => entry.process);
-}
-
-function processAttentionRank(process: ThreadnoteProcessDiagnostic): number {
-  if (process.role === 'legacy') return 2;
-  if (
-    process.activityRole !== undefined ||
-    (process.currentOperation !== undefined && !isProcessBaselineOperation(process))
-  ) {
-    return 0;
-  }
-  return 1;
-}
-
-function isProcessBaselineOperation(process: ThreadnoteProcessDiagnostic): boolean {
-  switch (process.role) {
-    case 'manager':
-      return process.currentOperation === 'manager-ui';
-    case 'mcp':
-      return process.currentOperation === 'mcp-server';
-    case 'mcp-broker':
-      return process.currentOperation === 'mcp-broker';
-    case 'local-model-worker':
-      return process.currentOperation === 'model-stdio';
-    case 'graph-parser-worker':
-      return process.currentOperation === 'parser-stdio';
-    default:
-      return false;
-  }
 }
 
 export type ThreadnoteProcessTerminationErrorCode =
